refactor(comic_builder): consolidate models and schemas into state/relational
Delete top-level models.py and schemas.py, moving Project model and
project CRUD schemas into state/relational/. Update all API imports
to reference the new locations. Extract SimpleEnvelope into its own
module and inline StoryPromptRequest into phases.py. Retain
backward-compatible default state initialization on the relational
Project model via initialize_empty_consolidated_state_dict.
