+++ backend/core/comic_builder/api/dependencies.py
@@ -1,19 +1,14 @@
 import uuid
 from typing import Annotated
 
-from fastapi import APIRouter, Depends, HTTPException
-from loguru import logger
+from fastapi import Depends, HTTPException
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from core.auth.dependencies import get_current_user_id
 from core.services.database import get_async_db_session
 
-from ..models import Project
-
-router = APIRouter(prefix="/phase", tags=["comic", "builder"])
-
-logger = logger.bind(name=__name__)
+from ..state.relational.models import Project
 
 
 async def verify_project_access(
