--- backend/core/comic_builder/models.py
@@ -1,35 +1,0 @@
-import uuid
-from datetime import datetime
-from typing import Any
-
-from sqlalchemy import DateTime, String
-from sqlalchemy.dialects.postgresql import JSONB, UUID
-from sqlalchemy.orm import Mapped, mapped_column
-from sqlalchemy.schema import ForeignKey
-
-from core.comic_builder.state import initialize_empty_consolidated_state_dict
-from core.common import ORMBase
-from core.common.utils import get_current_datetime_utc
-
-
-class Project(ORMBase):
-    __tablename__ = "project"
-
-    id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), primary_key=True, default=uuid.uuid4
-    )
-    user_id: Mapped[uuid.UUID] = mapped_column(
-        UUID(as_uuid=True), ForeignKey("user.id"), nullable=False
-    )
-    name: Mapped[str] = mapped_column(String(255), nullable=True)
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), default=get_current_datetime_utc
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True),
-        default=get_current_datetime_utc,
-        onupdate=get_current_datetime_utc,
-    )
-    state: Mapped[dict[str, Any]] = mapped_column(
-        JSONB, default=initialize_empty_consolidated_state_dict
-    )
